feat(imobiliaria api): adicionar rota para criação de uma reserva

+++ modulo-6-dotNet/TrabalhoFinal/Back/Imobiliaria/Imobiliaria.Dominio/Entidades/Opcional.cs
@@ -10,7 +10,7 @@ namespace Imobiliaria.Dominio.Entidades
     {
         public int Id { get; set; }
         public string Descricao { get; set; }
-        public decimal Preco { get; set; }
+        public int Preco { get; set; }
         public int Quantidade { get; set; }
     }
 }
