+++ modulo-6-dotNet/TrabalhoFinal/Back/Imobiliaria/Imobiliaria.Dominio/Entidades/Pacote.cs
@@ -11,6 +11,6 @@ namespace Imobiliaria.Dominio.Entidades
         public int Id { get; set; }
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public decimal Preco { get; set; }
+        public int Preco { get; set; }
     }
 }
